Workaround to get HDMI working
Xen doesn't map any range of memory not explicitly specified in the device
tree. Unfortunately, the framebuffer memory region is not specified in the
device tree, so Xen doesn't map that region. Accesses to the framebuffer memory
were being trapped by Xen, leading to a Linux kernel panic.

Work around this by introducing a dummy device to get Xen to map the frame
buffer region.

Signed-off-by: Stewart Hildebrand <stewart.hildebrand@dornerworks.com>